+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - Change the GRUB menu to show winesapOS as the distribution name.
 - Add winesapOS logo to the login screen.
 - Change Snapper to only take 6 monthly snapshots of "/home".
+- Add Btrfs quotas for snapshots.
 
 ## 4.0.0 - The Conversion Update